Minifying apk size

+++ BingBin/app/src/main/java/io/bingbin/bingbinandroid/tensorflow/env/ImageUtils.java
@@ -136,9 +136,9 @@ public class ImageUtils {
         // nG = (int)(1.164 * nY - 0.813 * nV - 0.391 * nU);
         // nB = (int)(1.164 * nY + 1.596 * nV);
 
-        int nR = (int) (1192 * nY + 1634 * nV);
-        int nG = (int) (1192 * nY - 833 * nV - 400 * nU);
-        int nB = (int) (1192 * nY + 2066 * nU);
+        int nR = 1192 * nY + 1634 * nV;
+        int nG = 1192 * nY - 833 * nV - 400 * nU;
+        int nB = 1192 * nY + 2066 * nU;
 
         nR = Math.min(kMaxChannelValue, Math.max(0, nR));
         nG = Math.min(kMaxChannelValue, Math.max(0, nG));
